server/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
python-multipart==0.0.18
python-dotenv==1.1.0
pdfplumber==0.11.0
langchain-google-genai>=2.1.0
langchain-core>=0.3.43
pydantic>=2.11.0
